feat(parser): add link conversor script for markdown and obsidian
This commit addresses an issue on the links and paths compatibility between the GitHub Flavored Markdown (GFM) and the Obsidian Markdown, by adding a shell script that does all the replacements required for each environment/parser. This script guarantees almost the same experience in both environments, enhancing the reading and writting for both environments.