fix(convert-file): ensure RegExp matches file names with path resolution
This commit addresses a specific issue where the RegExp was only checking for filenames and not checking the path resolving that name. This caused a mismatch on all repeated README filenames even those were in different folders. To solve this issue i follow those links until resolution to then match the changed file and updating only the raw matching if the route complies.